requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
cachetools
pydantic
pyjwt
redis